rename types.py to specs.py
