+++ libensemble/libE.py
@@ -177,7 +177,7 @@ from libensemble.executors.executor import Executor
 from libensemble.resources.resources import Resources
 from libensemble.tools.tools import _USER_SIM_ID_WARNING
 from libensemble.tools.alloc_support import AllocSupport
-from libensemble.types import SimSpecs, GenSpecs, AllocSpecs, ExitCriteria, LibeSpecs, EnsembleSpecs
+from libensemble.specs import SimSpecs, GenSpecs, AllocSpecs, ExitCriteria, LibeSpecs, EnsembleSpecs
 
 logger = logging.getLogger(__name__)
 # To change logging level for just this module
