+++ libensemble/tests/unit_tests/test_models.py
@@ -2,7 +2,7 @@ import numpy as np
 
 import libensemble.tests.unit_tests.setup as setup
 from libensemble.tools.fields_keys import libE_fields
-from libensemble.types import SimSpecs, GenSpecs, ExitCriteria, LibeSpecs, EnsembleSpecs
+from libensemble.specs import SimSpecs, GenSpecs, ExitCriteria, LibeSpecs, EnsembleSpecs
 
 
 class Fake_MPI:
